New IsoWorld system to test the rendering system.
- Added a new Box3 -> Box3Sized
- Bug correction in Mouse Wheel.
- Bug correction in Mouse Drag, when going out of the window.
- Library : renamed getOrAdd into addAndGet
- Library : added getOrAdd with a code block to generate the Resource.
- Added a Zoom avatar event (maps to scroll wheel or pinch gestures).
- Added a way to pass initialization code in Renderer.start() to avoir
thread problems.
- Added a mesh.draw(gl) method in Mesh to avoid using the last vertex
array, shortcut to mesh.lastva.draw(mesh.drawAs(gl)).

+++ android/src/org/sofa/opengl/backend/SurfaceAndroidES20.scala
@@ -154,7 +154,7 @@ class SurfaceAndroidES20(context:Context, attrs:android.util.AttributeSet)
 
 class ScrollEventAndroid(val source:android.view.MotionEvent) extends org.sofa.opengl.surface.ScrollEvent {
     // Not supported in > 10 (> android 2.3)
-    def amount:Int = 0//source.getAxisValue(android.view.MotionEvent.AXIS_VSCROLL).toInt
+    def amount:Double = 0//source.getAxisValue(android.view.MotionEvent.AXIS_VSCROLL).toInt
 }
 
 class MotionEventAndroid(val source:android.view.MotionEvent) extends org.sofa.opengl.surface.MotionEvent {
